Improve resiliency of the GithubSearchTest
This change also changed the way annotations like @Scenario and @BDD4jTest are used. @BDD4jTest is now a meta annotation and @Scenario is annotated with it. As @Scenario is now also meta annotated with the @Test annotation, only the @Scenario annotation has to be present when defining a test scenario.

This commit also changes a few of the internals as to how the framework processes the TestState and improves resource handling.

https://github.com/bdd4j/bdd4j/issues/8