Replace BDD4jRunner with BDD4jScenario
To improve the design of the libraries internals the static BDD4jRunner#scenario method has been replaced with an actual BDD4jScenario class.

However, this has one major drawback. To allow type safe definitions of steps the TestState needs to be defined as a generic type in the definition of the ScenarioBuilder class. Therefor injecting the steps is now a bit more chatty than it was before.

https://github.com/bdd4j/bdd4j/issues/30