Add TableParser and DataRow classes with tests
Implemented the TableParser class for parsing cucumber data tables and DataRow class for representing parsed rows. Also, added a TableParserTest class to test the functionality of the TableParser. DataRow class includes methods for fetching the string from a given key as well as equal and hash functions. TableParser considers the first line as column keys, splitting on vertical bars and storing values in a DataRow instance.